refactor: Models.cs をクラスごとにファイル分割 (#191)
ExtensionInfo / TimelineConfig / ProfileConfig / AppSettings を
1 クラス 1 ファイルに分割。名前空間は変更なし。
テストプロジェクトの参照はワイルドカード（Models\*.cs）に変更。
