+++ Models/ExtensionInfo.cs
@@ -1,0 +1,11 @@
+namespace XTimelineViewer.Models
+{
+    internal record ExtensionInfo(
+        string Name,
+        string DirectoryPath,
+        string? IconPath,
+        string? OptionsPage,
+        string? HomepageUrl,
+        string? ExtensionId
+    );
+}
